Move PortableRegistry to own file, introduce PortableRegistryBuilder
