+++ src/tests.rs
@@ -24,7 +24,6 @@ use crate::{
         string::String,
         vec,
     },
-    registry::PortableType,
     *,
 };
 use core::marker::PhantomData;
@@ -363,41 +362,4 @@ fn basic_enum_with_index() {
         );
 
     assert_type!(IndexedRustEnum, ty);
-}
-
-#[test]
-fn construct_portable_registry() {
-    let mut types = Vec::new();
-
-    let u32_type_id = types.len() as u32;
-    types.push(PortableType::new(
-        u32_type_id,
-        Type::new(Path::default(), vec![], TypeDefPrimitive::U32, vec![]),
-    ));
-
-    let vec_u32_type_id = types.len() as u32;
-    types.push(PortableType::new(
-        vec_u32_type_id,
-        Type::new(
-            Path::default(),
-            vec![],
-            TypeDefSequence::new(u32_type_id.into()),
-            vec![],
-        ),
-    ));
-
-    let composite_type_id = types.len() as u32;
-    let composite = Type::builder_portable()
-        .path(Path::new_custom(vec!["MyStruct".into()]))
-        .composite(
-            Fields::named()
-                .field_portable(|f| f.name("primitive".into()).ty(u32_type_id))
-                .field_portable(|f| f.name("vec_of_u32".into()).ty(vec_u32_type_id))
-                .field_portable(|f| {
-                    f.name("self_referential".into()).ty(composite_type_id)
-                }),
-        );
-    types.push(PortableType::new(composite_type_id, composite));
-
-    let _registry = PortableRegistry::new_from_types(types);
 }
